Revert "By default, enable built-in fonts in configure script."
This reverts commit f7b940677eae414ab08ce41fe6bdb7253b7a7864.

The built-in fonts (--enable-builtin-fonts) trigger a segfault with
clients. That is, Xprt will build and run successfully with them, but
when a client such as xpsimplehelloworld tries to print via Xprt, it
segfaults.  Since Xprint is deprecated, motivation is not high for
fully debugging this problem.

The workaround around will be to not use built-in fonts (i.e.
reverting the previous commit to use --disable-builtin-fonts by
default again), and build against libXfont 1.3 (rather than the latest
libXfont 1.4, which has Xprint support removed).